+++ settings.py
@@ -23,7 +23,7 @@ APP_KITS = {
             'task1_results_tr12',
             'task2_instructions',
             'task2_prac_tr13',
-            'new_task2_quiz',
+            'task2_quiz',
             'task2_tr1',
             'new_task2_results',
             'static_portfolio',
@@ -46,7 +46,7 @@ APP_KITS = {
             'task1_results_tr12',
             'task2_instructions',
             'task2_prac_tr2',
-            'new_task2_quiz',
+            'task2_quiz',
             'task2_tr2',
             'new_task2_results',
             'static_portfolio_chp',
@@ -69,7 +69,7 @@ APP_KITS = {
             'tr3_dynamic_pooled_results',
             'task2_instructions',
             'task2_prac_tr13',
-            'new_task2_quiz',
+            'task2_quiz',
             'new_task2',
             'new_task2_results',
             'static_portfolio',
@@ -103,7 +103,7 @@ SESSION_CONFIGS = [
                       'task2_instructions',
                       'task2_prac_tr13',
                       'task2_prac_tr2',
-                      'new_task2_quiz',
+                      'task2_quiz',
                       'task2_tr1',
                       'task2_tr2',
                       'new_task2_results',
