+++ task2_quiz/models.py
@@ -11,7 +11,7 @@ Your app description
 
 
 class Constants(BaseConstants):
-    name_in_url = 'new_task2_quiz'
+    name_in_url = 'task2_quiz'
     players_per_group = None
     num_rounds = 1
     initial_wealth = 100
