+++ task2_quiz/pages.py
