+++ task2_quiz/tests.py
@@ -1,0 +1,9 @@
+from otree.api import Currency as c, currency_range
+from . import pages
+from ._builtin import Bot
+from .models import Constants
+
+
+class PlayerBot(Bot):
+    def play_round(self):
+        pass
